requirements.txt
aspose-words==26.3.0
